requirements.txt
# Local development — ingest scripts
PyMuPDF>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
rank-bm25==0.2.2
numpy>=1.26.0
